task5/requirements.txt
pytest
appium-python-client
selenium
